smartcity/requirements.txt
fastapi
uvicorn[standard]
pydgraph
grpcio
pymongo
cassandra-driver